adicionado botao de localizaçao no perfil

+++ screens/ProfileScreen.tsx
@@ -9,6 +9,9 @@ import {
 } from 'firebase/auth';
 import { useAuth } from '../context/AuthContext';
 import { ProfileForm } from '../components/ProfileForm';
+import * as Location from 'expo-location';
+import { Linking } from 'react-native';
+
 
 export default function ProfileScreen() {
   const [user, setUser] = useState(auth.currentUser);
@@ -17,6 +20,7 @@ export default function ProfileScreen() {
   const [photoUrl, setPhotoUrl] = useState('');
   const [password, setPassword] = useState('');
   const { isLoggedIn, setIsLoggedIn } = useAuth();
+  const [userLocation, setUserLocation] = useState(null);
 
   if (email === null) {
     return;
@@ -38,11 +42,13 @@ export default function ProfileScreen() {
         setName(data.name);
         setEmail(data.email);
         setPhotoUrl(data.photoUrl);
+        setUserLocation(data.userLocation);
       } else {
         await setDoc(userProfileDoc, {
           name: user.displayName || '',
           email: user.email,
           photoUrl: user.photoUrl || '',
+          userLocation: user.userLocation || null,
         });
       }
     }
@@ -67,6 +73,7 @@ export default function ProfileScreen() {
           name,
           email,
           photoUrl,
+          userLocation,
         },
         { merge: true }
       );
@@ -91,27 +98,58 @@ export default function ProfileScreen() {
 
   const handleDeleteAccount = async () => {
     if (user) {
-      try {
-        const userProfileDoc = doc(db, 'users', user.uid);
-        await deleteDoc(userProfileDoc);
-        await deleteUser(user);
-        Alert.alert('Conta deletada', 'Sua conta foi deletada com sucesso.');
-        setIsLoggedIn(false);
-        setUser(null);
-        setName('');
-        setEmail('');
-        setPassword('');
-        setPhotoUrl('');
-      } catch (error: any) {
-        console.error('Erro ao deletar conta: ', error);
-        Alert.alert(
-          'Erro',
-          'Ocorreu um erro ao deletar sua conta. Por favor, tente novamente.'
-        );
-      }
+      Alert.alert(
+        'Confirmação',
+        'Tem certeza que deseja deletar sua conta? Esta ação não pode ser desfeita.',
+        [
+          {
+            text: 'Cancelar',
+            onPress: () => console.log('Cancelado'),
+            style: 'cancel',
+          },
+          {
+            text: 'Sim',
+            onPress: async () => {
+              try {
+                const userProfileDoc = doc(db, 'users', user.uid);
+                await deleteDoc(userProfileDoc);
+                await deleteUser(user);
+  
+                Alert.alert('Conta deletada', 'Sua conta foi deletada com sucesso.');
+                setIsLoggedIn(false);
+                setUser(null);
+                setName('');
+                setEmail('');
+                setPassword('');
+                setPhotoUrl('');
+                handleLogout();
+              } catch (error: any) {
+                console.error('Erro ao deletar conta: ', error);
+                Alert.alert(
+                  'Erro',
+                  'Ocorreu um erro ao deletar sua conta. Por favor, tente novamente.'
+                );
+              }
+            },
+          },
+        ],
+        { cancelable: false } // Impede que o alerta seja fechado ao clicar fora dele
+      );
     } else {
       Alert.alert('Erro', 'Nenhum usuário autenticado para deletar.');
     }
+  };
+
+  const getLocation = async () => {
+    let { status } = await Location.requestForegroundPermissionsAsync();
+    if (status !== 'granted') {
+      Alert.alert('Permissão para acessar a localização negada');
+      return;
+    }
+
+    let location = await Location.getCurrentPositionAsync({});
+    setUserLocation(location);
+    Alert.alert('Localização capturada', `Lat: ${location.coords.latitude}, Long: ${location.coords.longitude}`);
   };
 
   return (
@@ -124,6 +162,7 @@ export default function ProfileScreen() {
       onSaveProfile={handleSaveProfile}
       onLogout={handleLogout}
       onDeleteAccount={handleDeleteAccount}
+      onButtonLocation={getLocation}
     />
   );
 };
